Task 2
I have used food name and taste as the two data values. An all Get method can access all the data sent and display it. The food name Get method, shows the food and its taste, as according to the name entered by the user. The food taste Get method, shows the food and its taste according to the taste entered by the user. The post method, allows you to enter the food name and its taste for using later on.